+++ CHANGELOG.md
@@ -7,6 +7,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.5] - 2026-03-08
+
+### Docs
+- Update project/README.md
+- Update project/context.md
+
+### Test
+- Update tests/project/dashboard.html
+- Update tests/project/project.yaml
+
+### Other
+- Update code2docs/cli.py
+- Update code2docs/config.py
+- Update code2docs/generators/_registry_adapters.py
+- Update code2docs/generators/org_readme_gen.py
+- Update project/analysis.json
+- Update project/analysis.toon
+- Update project/analysis.yaml
+- Update project/calls.mmd
+- Update project/compact_flow.mmd
+- Update project/dashboard.html
+- ... and 7 more files
+
 ## [3.0.4] - 2026-03-08
 
 ### Docs